Added first simple CUDA  sample
This is a simple CUDA sample that shows how to run a kernel that has the basic structure that will be used later to build a ray tracer. The output shown is a normalized 0-1 coordinate space. More changes might be applied.
I've added a simple *.bmp saving function utility  that is used by the sampl itself to save the buffer results in *.bmp format.