Fixed synchronization bug with fences.
I've added a FlushGPU() helper function and a WaitForPreviousFrame(). They are both needed. The first to ensure that any command list in any of the 3 backbuffer has finished executiion and the second just to check fence related to ensure previous frame cmd list has completed its work. Either way wait for previous frame to complete will stall the CPU therefore is not the best approach efficency-wise